requirements.txt
fastapi
uvicorn
openai-whisper
torch
torchvision
torchaudio
ffmpeg-python
numpy
faster-whisper